+++ RefactorThis.Persistence/Models/Payment.cs
@@ -1,0 +1,8 @@
+namespace RefactorThis.Persistence.Models
+{
+    public class Payment
+    {
+        public decimal Amount { get; set; }
+        public string Reference { get; set; }
+    }
+}
