--- RefactorThis.Persistence/Payment.cs
@@ -1,8 +1,0 @@
-namespace RefactorThis.Persistence
-{
-	public class Payment
-	{
-		public decimal Amount { get; set; }
-		public string Reference { get; set; }
-	}
-}
