+++ RefactorThis.Persistence/RefactorThis.Persistence.csproj
@@ -5,4 +5,8 @@
 		<LangVersion>7.3</LangVersion>
 	</PropertyGroup>
 
+	<ItemGroup>
+	  <Folder Include="Interface\" />
+	</ItemGroup>
+
 </Project>
